fibonacci 2

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,9 +8,9 @@
 int main(void)
 {
 int count;
-unsigned long int fib1 = 1, fib2 = 2, next_fib;
+unsigned long long int fib1 = 1, fib2 = 2, next_fib;
 
-printf("%lu, %lu, ", fib1, fib2);
+printf("%llu, %llu, ", fib1, fib2);
 
 for (count = 3; count <= 98; count++)
 {
@@ -19,9 +19,9 @@ fib1 = fib2;
 fib2 = next_fib;
 
 if (count != 98)
-printf("%lu, ", next_fib);
+printf("%llu, ", next_fib);
 else
-printf("%lu\n", next_fib);
+printf("%llu\n", next_fib);
 }
 
 return (0);
